+++ dbgr/session.py
@@ -1,3 +1,4 @@
+from types import SimpleNamespace
 import aiohttp
 from dbgr.reporting import ProgressBar, Reporter
 
@@ -5,12 +6,18 @@ from dbgr.reporting import ProgressBar, Reporter
 _SESSION = None
 
 
+class Session(aiohttp.ClientSession):
+    async def _request(self, method, url, **kwargs): #pylint: disable=W0221
+        kwargs['trace_request_ctx'] = SimpleNamespace(method=method, url=url, **kwargs)
+        return await super()._request(method, url, **kwargs)
+
+
 def get_session():
     global _SESSION # pylint: disable=W0603
     if not _SESSION:
         progress_bar = ProgressBar()
         reporter = Reporter()
-        _SESSION = aiohttp.ClientSession(
+        _SESSION = Session(
             trace_configs=[
                 progress_bar.get_tracer(),
                 reporter.get_tracer()